feat(nixec): have eka act as nixec when arg0 set
eka will bootstrap evaluations by reexecing itself as nixec and running
the sandboxed evaluations.
